fix #5, support Fitt's law
--the buttons have gained paddings in order
to be used to accept mouse hovering in that
areas. When the user has set the length margin
which is near the screen edge to zero the button
located there it will always accept mouse
events event when the mouse touches the far
edge of the screen. This is why all length
margins have been set to zero for default
values.