fix #63,use a common shared decoration for buttons
--a new class has been introduced called PreviewSharedDecoration
and is the common decoration that is used for all PreviewButtons.
This way more resources are gained and at the same time the buttons
can communicate with each other through the decoration if it
is needed from decoration perspective. This was not straight forward
as the coordinates between KDecoration2::Button(s) and their
QML parts must be translated in a different way in order to work
correctly.